Implement sliding window rate limiting with Redis support and add Lua script for execution

+++ src/rate_limiter/request.py
@@ -18,5 +18,5 @@ class Client(BaseModel):
 
 
 class Rules(BaseModel):
-    max_requests: float
-    time_window: float  # in seconds
+    max_requests: int
+    time_window: int  # in seconds
